+++ community/cypher/cypher-compiler-2.2/src/main/scala/org/neo4j/cypher/internal/compiler/v2_2/pipes/SemiApplyPipe.scala
@@ -25,7 +25,7 @@ import org.neo4j.cypher.internal.compiler.v2_2.planDescription.{PlanDescriptionI
 import org.neo4j.cypher.internal.compiler.v2_2.symbols.SymbolTable
 
 case class SemiApplyPipe(source: Pipe, inner: Pipe, negated: Boolean)
-                        (val estimatedCardinality: Option[Long] = None)(implicit pipeMonitor: PipeMonitor)
+                        (val estimatedCardinality: Option[Double] = None)(implicit pipeMonitor: PipeMonitor)
   extends PipeWithSource(source, pipeMonitor) with RonjaPipe {
   def internalCreateResults(input: Iterator[ExecutionContext], state: QueryState): Iterator[ExecutionContext] = {
     input.filter {
@@ -52,5 +52,5 @@ case class SemiApplyPipe(source: Pipe, inner: Pipe, negated: Boolean)
 
   override def localEffects = Effects.NONE
 
-  def withEstimatedCardinality(estimated: Long) = copy()(Some(estimated))
+  def withEstimatedCardinality(estimated: Double) = copy()(Some(estimated))
 }
