+++ community/cypher/cypher-compiler-2.2/src/main/scala/org/neo4j/cypher/internal/compiler/v2_2/planDescription/PlanDescriptionArgumentSerializer.scala
@@ -40,7 +40,7 @@ object PlanDescriptionArgumentSerializer {
       case DbHits(value) => Long.box(value)
       case _: EntityByIdRhs => arg.toString
       case Rows(value) => Long.box(value)
-      case EstimatedRows(value) => Long.box(value)
+      case EstimatedRows(value) => Double.box(value)
       case Version(version) => version
       case Planner(planner) => planner
       case ExpandExpression(from, rel, typeNames, to, dir: Direction, varLength) =>
